requirements.txt
streamlit
pypdf
mammoth
python-docx
tiktoken
openai>=1.0.0